% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_add_ffdi}
\alias{bom_db_add_ffdi}
\title{Add values for KBDI, Drought Factor and FFDI to database records}
\usage{
bom_db_add_ffdi(
  db,
  tables = c("synoptic", "aws"),
  states = c("NSW", "VIC", "ACT", "TAS", "QLD", "SA", "WA"),
  cutoff = as.Date("2000-01-01"),
  recalculate = FALSE
)
}
\arguments{
\item{db}{A database connection object created with \code{pool::dbPool} or
\code{DBI::dbConnect}.}

\item{tables}{The tables to process. One or both of \code{'aws'} and \code{'synoptic'}.
Default is both.}

\item{states}{Standard abbreviations for the states to process. Default is
\code{c('NSW', 'VIC', 'ACT', 'TAS', 'QLD', 'SA', 'WA')}.}

\item{cutoff}{Cutoff date to subset stations to process. Values will only be
calculated for stations that were active (i.e. have records in the
database) on or after this date. Can be specified as a \code{Date} object
or a character string in \code{'yyyy-mm-dd'} format. Default is
\code{'2000-01-01'}.}

\item{recalculate}{If \code{TRUE}, values will be calculated for all records and any existing values
will be over-written. If \code{FALSE} (default), values will only be calculated for the most recent
records that do not already have FFDI values.}
}
\description{
This function uses the \code{CERMBffdi} package to calculate values for the
fire related variables: KBDI, drought factor and FFDI for selected weather
records, and store the values in the corresponding database columns. With
default arguments, the function will process all weather stations in the
'aws' and 'synoptic' tables that were active on or after 2000-01-01. The
function attempts to minimize data transfer over the network and running time
by only processing the most recent records that do not already have FFDI
values. Set the argument \code{recalculate=TRUE} to override this behaviour
and force recalculation for all records.
}
