% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_functions.R
\name{bom_db_calculate_ffdi}
\alias{bom_db_calculate_ffdi}
\title{Calculate FFDI for a given data set.}
\usage{
bom_db_calculate_ffdi(
  dat,
  av.rainfall = NULL,
  datatype = c("guess", "AWS", "Synoptic"),
  windcol = NULL
)
}
\arguments{
\item{dat}{A data frame with records for one or more weather stations.
Normally this will made up of records returned from querying the synoptic
or AWS tables in the database. If from another data source, the columns
and column names must conform to those used in the database tables. The
required columns are:
\code{'station', 'year', 'month', 'day', 'hour', 'minute',
      'precipitation', 'temperature', 'relhumidity'}
plus one or both of \code{'windspeed'} or \code{'windgust'} (AWS data).}

\item{av.rainfall}{Average annual rainfall value to use in the calculation of
KBDI (on which FFDI relies). Note: if the data include multiple weather
stations, the same average rainfall value is applied to all.}

\item{datatype}{Either a specified data type ('AWS' or 'Synoptic') or 'guess'
(default) to guess the type from the data. Case-insensitive and may be
abbreviated. The data type controls how rainfall values are aggregated to
daily values for the calculation of KBDI and drought factor. AWS rainfall
values are total since 9am, whereas Synoptic values are rainfall for the
record time step. If 'guess', the function checks whether non-zero rainfall
values are always ascending during each 24 hour period to 9am which
indicates AWS data.}

\item{windcol}{Name of the column of wind speed values to use for FFDI
calculation. The default is to use \code{'windgust'} if present, otherwise
\code{'windspeed'}.}
}
\value{
A data frame with columns:
  \code{'station', 'year', 'month', 'day', 'hour', 'minute',}
  \code{'tmaxdaily'} (maximum temperature for calendar day),
  \code{'precipdaily'} (total rainfall from 09:01 to 09:00 the next day),
  \code{'kbdi'} (Keetch-Bryam drought index),
  \code{'drought'} (daily drought factor values),
  \code{'ffdi'} (FFDI value based on daily drought factor and time-step
    values for temperature, wind speed and relative humidity.
}
\description{
This function calculates FFDI (Forest Fire Danger Index)
and the related variables: KBDI (Keetch-Bryam Drought Index) and Drought Factor.
You can call it directly to calculate FFDI for an arbitrary data set of
weather data for one or more stations (see Arguments below for the required
format). The function is also used by \code{bom_db_update_fire} which sets the
values of fire-related variables in the database tables.
}
\examples{
\dontrun{
library(CERMBweather)
library(dplyr, warn.conflicts = FALSE)

# Get synoptic data for selected weather stations
DB <- bom_db_open("/path/to/database/BOM.db")

stns <- c(1006, 1007, 1019, 1020)

dat <- bom_db_synoptic(DB) \%>\%
  filter(station \%in\% stns) \%>\%
  collect()

# Calculate FFDI for stations using a specified value for
# average annual rainfall
ffdi <- bom_db_calculate_ffdi(dat, av.rainfall.value = 800)

# Calculate FFDI for stations, with average rainfall being
# calculated over the default reference period of 2001-2015
ffdi <- bom_db_calculate_ffdi(dat)

# Calculate FFDI for stations using a non-default period
# for the calculation of average annual rainfall
ffdi <- bom_db_calculate_ffdi(dat, av.rainfall.years = 2016:2018)

# If you want to relate FFDI and drought factor values to the
# original sub-daily weather values, join to input data
ffdi <- left_join(dat, ffdi,
                  by = c("station", "year", "month", "day", "hour", "minute"))

bom_db_close(DB)
}

}
\seealso{
\code{\link{bom_db_update_fire}}
}
