% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_check_datetimes}
\alias{bom_db_check_datetimes}
\title{Check that records for a given station form an uninterrupted time series}
\usage{
bom_db_check_datetimes(db, station, dbtable = c("aws", "synoptic"))
}
\arguments{
\item{db}{A database connection pool object as returned by
\code{\link{bom_db_open}} or \code{\link{bom_db_create}}.}

\item{station}{Integer station identifier.}

\item{dbtable}{Either 'AWS' or 'Synoptic'. Case-insensitive and may be
abbreviated. Default is 'AWS'.}
}
\value{
A named list with elements:
  \describe{
    \item{ok}{TRUE if records form an uninterrupted time series; FALSE otherwise.}
    \item{err}{If 'ok' is FALSE, message indicating first break in time series.}
  }
}
\description{
Checks whether the records for each station form a non-interrupted time series.
}
