% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_functions.R
\name{bom_db_ffdi}
\alias{bom_db_ffdi}
\title{Calculate FFDI for specified weather stations}
\usage{
bom_db_ffdi(dat, av.rainfall.years = 2001:2015, min.days.per.year = 300)
}
\arguments{
\item{av.rainfall.years}{A vector of four digit year numbers for the
reference period over which average annual rainfall at each weather station
should be calculated. This is used in the KBDI calculation. Normally the
reference period will be a sequence of contiguous years, but it does not
have to be. Default is \code{2001:2015}.}

\item{min.days.per.year}{The minimum number of days of rainfall data required
for each year specified by \code{av.rainfall.years}. Default is \code{300}.}

\item{dat.stn}{A data frame with records for one or more weather stations.
Normally this will made up of records returned from querying the synoptic
or AWS tables in the database. If from another data source, the columns
and column names must conform to those used in the database tables for
variables used to calculate KBDI and FFDI values:
\code{'station', 'year', 'month', 'day', 'hour', 'minute',
      'precipitation', 'temperature', 'relhumidity', 'windspeed'}}
}
\value{
A data frame with columns:
  \code{'station', 'year', 'month', 'day', 'hour', 'minute',}
  \code{'tmax.daily'} (maximum temperature for calendar day),
  \code{'precip.daily'} (total rainfall from 09:01 to 09:00 the next day),
  \code{'drought'} (daily drought factor values),
  \code{'ffdi'} (FFDI value based on daily drought factor and time-step
  values for temperature, wind speed and relative humidity.
}
\description{
Calculate FFDI for specified weather stations
}
\examples{
\dontrun{
library(CERMBweather)
library(dplyr, warn.conflicts = FALSE)

# Get synoptic data for selected weather stations
DB <- bom_db_open("/path/to/database/BOM.db")

stns <- c(1006, 1007, 1019, 1020)

dat <- bom_db_synoptic(DB) \%>\%
  filter(station \%in\% stns) \%>\%
  collect()

# Calculate FFDI for stations
ffdi <- bom_db_ffdi(dat)

# If you want to relate FFDI and drought factor values to the
# original sub-daily weather values, join to input data
ffdi <- left_join(dat, ffdi,
                  by = c("station", "year", "month", "day", "hour", "minute"))

bom_db_close(DB)
}

}
