% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_functions.R
\name{bom_db_get_daily_rainfall}
\alias{bom_db_get_daily_rainfall}
\title{Get daily aggregate rainfall data}
\usage{
bom_db_get_daily_rainfall(
  db,
  the.table,
  stations,
  crop = TRUE,
  start.date = NULL,
  end.date = NULL,
  dry.run = FALSE
)
}
\arguments{
\item{db}{A database connection pool object.}

\item{the.table}{The table to query: either 'aws' or 'synoptic'
(may be abbreviated).}

\item{stations}{A vector containing one or more BOM weather station integer
IDs. Any values not present in the table will be silently ignored so, for
example, you could retrieve data for all (almost) NSW stations by calling
the function with \code{stations = 46000:75999}.}

\item{crop}{If \code{TRUE} (default), the returned values will not include
the earliest rain date (09:01 - 09:00) in the series as this is most
probably a partial day.}

\item{start.date}{The start date (local time) for the query, provided as
either a \code{Date} object or a character string in the format
\code{yyyy-mm-dd}. The default is to start from the earliest data for
each station.}

\item{end.date}{The end date (local time) for the query, provided as
either a \code{Date} object or a character string in the format
\code{yyyy-mm-dd}. The default is to end with the latest data for each
station.}

\item{dry.run}{If \code{TRUE}, the function returns the SQL code for the
query but does not run it. This can be useful if you want to further
modify the query. If \code{FALSE} (default), the function runs the query
and returns the resulting data.}
}
\value{
A data frame with columns: station, date_rain (09:01 - 09:00 local
  time), precip_daily.
}
\description{
Get daily aggregate rainfall data
}
