% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_import}
\alias{bom_db_import}
\title{Import weather station data from a zip file as provided by BOM}
\usage{
bom_db_import(con, zipfile, stations = NULL, allow.missing = TRUE,
  progress = FALSE)
}
\arguments{
\item{con}{An open database connection as returned by \code{\link{bom_db_init}}.}

\item{zipfile}{Path to the input zip file.}

\item{stations}{Either NULL (default) to import data for all stations from
the zip file; or a character vector of station identifiers.}

\item{allow.missing}{If TRUE (default) and specific stations were requested,
the function will silently ignore any that are missing in the zip file. If
FALSE, missing stations result in an error.}

\item{progress}{If TRUE display a progress bar during the import.
If FALSE (default), do not display a progress bar.}
}
\value{
A named vector giving the number of rows added to the AWS and Synoptic
  tables.
}
\description{
Reads weather data from a zip file, in the format provided by the Bureau of
Meteorology, containing data sets for one or more weather stations in CSV
format, and imports them into a SQLite database. Any duplicate records are
silently ignored.
}
