% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_init}
\alias{bom_db_init}
\title{Opens or initializes a database for weather data}
\usage{
bom_db_init(db, existing = FALSE, copy = FALSE)
}
\arguments{
\item{db}{Either an open connection to a SQLite database; a character
path to an existing database to open; or a character path to a
database to create (if argument \code{existing} is FALSE).}

\item{existing}{If TRUE and \code{db} is a path string then an error will
be thrown if the database file does not already exist. If FALSE (default),
the file will be created if necessary. Ignored if \code{db} is a
connection object rather than a path string.}

\item{copy}{If TRUE and \code{db} is an open database connection, the function
returns a copy of the connection. If FALSE, the returned object will be
the same as the input connection. Ignored if \code{db} is a path string.}
}
\value{
A connection (\code{SQLiteConnection} object) to the database.
}
\description{
This function can be used to open an existing database or create a new
database with tables for synoptic and time series data. In the case of an
existing database, the function will create tables for synoptic and AWS data
if they do not already exist.
}
\examples{
\dontrun{
# Open a database file, or create it if it does not exist, and
# initialize the weather data tables as required.
con <- bom_db_init("c:/foo/bar/weather.db")

# Open an existing database file. An error is thrown if it does
# not exist.
con <- bom_db_init("c:/foo/bar/weather.db", existing = TRUE)

# Pass in an open connection to a database so that
# weather data tables will be created if not already present
con <- bom_db_init(con)
}

}
