% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_readonly}
\alias{bom_db_readonly}
\title{Check if a database connection is read-only}
\usage{
bom_db_readonly(db)
}
\arguments{
\item{db}{A database connection pool object as returned by
\code{\link{bom_db_open}} or \code{\link{bom_db_create}}.}
}
\value{
TRUE if the connection is read-only; or FALSE if it supports write
  operations.
}
\description{
Checks if the given database connection is read-only, ie. can be used to
retrieve data but not import new data. An error is issued if \code{db}
is not a currently open connection to a weather database.
}
\examples{
\dontrun{
DB <- bom_db_open("c:/foo/bar/weather.db")

# This will return TRUE
bom_db_readonly(DB)

# Re-open connection with write permission
bom_db_close(DB)
DB <- bom_db_open("c:/foo/bar/weather.db", readonly = FALSE)

# Now this will return FALSE
bom_db_readonly(DB)
}

}
