% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_readwrite}
\alias{bom_db_readwrite}
\title{Check if a database connection supports read and write operations}
\usage{
bom_db_readwrite(db)
}
\arguments{
\item{db}{A database connection pool object as returned by
\code{\link{bom_db_open}} or \code{\link{bom_db_create}}.}
}
\value{
TRUE if the connection supports write operations; or FALSE if it is
  read-only.
}
\description{
Checks if the given database connection can be used for write operations
(importing new data) as well as read operations (querying and retrieving
data). An error is issued if \code{db} is not a currently open connection to
a weather database.
}
\examples{
\dontrun{
DB <- bom_db_open("c:/foo/bar/weather.db")

# This will return FALSE
bom_db_readwrite(DB)

# Re-open connection with write permission
bom_db_close(DB)
DB <- bom_db_open("c:/foo/bar/weather.db", readonly = FALSE)

# Now this will return TRUE
bom_db_readwrite(DB)
}

}
