% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_tbl}
\alias{bom_db_tbl}
\title{Gets a `tbl` object for use with dplyr functions}
\usage{
bom_db_tbl(db, tblname = c("synoptic", "aws"))
}
\arguments{
\item{db}{Either an open connection to a SQLite database; or a character
path to an existing database to open.}
}
\description{
Gets a `tbl` object for use with dplyr functions
}
