% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_functions.R
\name{bom_db_update_fire}
\alias{bom_db_update_fire}
\title{Calculate values of fire-related variables in the database tables}
\usage{
bom_db_update_fire(
  db,
  tables = c("AWS", "Synoptic"),
  records = c("new", "all"),
  stations = NULL
)
}
\arguments{
\item{db}{A database connection pool object as returned by
\code{\link{bom_db_open}} or \code{\link{bom_db_create}}.}

\item{tables}{Names of tables to update. Default is to update both the
'AWS' and 'Synoptic' tables. Table names may be abbreviated.}

\item{records}{The records to update: either \code{'new'} (default) to only
update records with more recent dates and times than the last record with
calculated values; or \code{'all'} to calculate values for all records in a
table, replacing any previously calculated values. Note that for both
options, values can only be calculated for records that form a continuous
time series.}

\item{stations}{An optional vector of integer station numbers. If provided,
updates will only be performed for these stations. The default is to
update all stations.}
}
\value{
A vector with named elements giving the number of records updated by
  table.
}
\description{
This function calculates and stores values for the database fields 'kbdi'
(Keetch-Bryam Drought Index), 'drought' (Drought Factor), and 'ffdi' (Forest
Fire Danger Index) for records in one or both of the 'AWS' and 'Synoptic'
database tables. Calculations are performed by the function
\code{bom_db_calculate_ffdi}. Depending on the value of the \code{records}
argument, values can be calculated for all records (any previous values are
replaced), or only for those records with more recent dates and times than
the last record with non-missing values. The records must cover a continuous
time series, i.e. no gaps in the sequence of dates.
}
\seealso{
\code{\link{bom_db_calculate_ffdi}}
}
