% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_functions.R
\name{bom_dir_summary}
\alias{bom_dir_summary}
\title{Summarizes weather station files in a directory}
\usage{
bom_dir_summary(path, include = c("data", "all", "empty"))
}
\arguments{
\item{path}{Path to the input directory.}

\item{include}{One of the following (may be abbreviated):
'data' (default) to only list stations with non-empty data sets;
'all' to list all stations;
'empty' to only list stations with empty data sets.}
}
\value{
A data frame with columns station (identifier); filename; filesize.
}
\description{
This function summarizes the contents of a directory containing delimited
text data files for one or more weather stations.
}
