% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_import}
\alias{bom_import}
\title{Import weather station data from a zip file}
\usage{
bom_import(zipfile, db, stations = NULL, allow.missing = TRUE)
}
\arguments{
\item{zipfile}{Path to the input zip file.}

\item{db}{Either an open database connection or the path to a database file to
create or update.}

\item{stations}{Either NULL (default) to import data for all stations from
the zip file; or a character vector of station identifiers.}

\item{allow.missing}{If TRUE (default) and specific stations were requested,
the function will silently ignore any that are missing in the zip file. If
FALSE, missing stations result in an error.}
}
\value{
Connection to the destination database.
}
\description{
Imports data from a zip file, containing data sets for one or more weather
stations in CSV format, into a SQLite database.
}
