% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_station_id}
\alias{bom_station_id}
\title{Format station identifying numbers as character strings}
\usage{
bom_station_id(id)
}
\arguments{
\item{id}{Vector (integer or character) of one or more station
identifying numbers.}
}
\value{
Vector of formatted identifiers.
}
\description{
Formats station numbers to match the embedded strings in station
data file names. Identifiers can be provided as integers or strings.
A valid station number has between four and six digits.
}
