% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipfile_functions.R
\name{bom_zip_data}
\alias{bom_zip_data}
\title{Reads raw data for specified weather stations from a zip file}
\usage{
bom_zip_data(
  zipfile,
  stations,
  allow.missing = TRUE,
  out.format = c("list", "single")
)
}
\arguments{
\item{zipfile}{Path to the input zip file.}

\item{allow.missing}{If TRUE (default), the function will return NULL if the
requested station is not found in the zip file. If FALSE, an error
is thrown if the station is missing.}

\item{out.format}{One of 'list' (default) to return a named list of data frames,
or 'single' to return a single frame of data for all stations.}

\item{station}{Station identifiers as integer station numbers or character strings.}
}
\value{
If out.format is 'list', a named list, where names are station
  identifiers (character strings of length six) and elements are data frames
  of station data. Each data frame has an attribute 'datatype' with value 'synoptic'
  or 'aws'. If out.format is 'single', a combined data frame of data for all
  stations, with an attribute 'datatype'.
}
\description{
Reads raw data for specified weather stations from a zip file
}
