% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipfile_functions.R
\name{bom_zip_summary}
\alias{bom_zip_summary}
\title{Summarizes weather stations in a zipped data file}
\usage{
bom_zip_summary(zipfile, include = c("data", "all", "empty"))
}
\arguments{
\item{zipfile}{Path to the input zip file.}

\item{include}{One of the following (may be abbreviated):
'data' (default) to only list stations with non-empty data sets;
'all' to list all stations;
'empty' to only list stations with empty data sets.}
}
\value{
A data frame with columns station (identifier); filename; filesize.
}
\description{
This function summarizes the contents of a zip file containing data sets
for one or more weather stations.
}
