% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_functions.R
\name{calculate_daily_kbdi}
\alias{calculate_daily_kbdi}
\title{Calculate the Keetch-Byram drought index (soil moisture deficit)}
\usage{
calculate_daily_kbdi(daily.precip, daily.tmax, average.rainfall)
}
\arguments{
\item{daily.precip}{A vector of daily (assumed sequential) precipitation
values.}

\item{daily.tmax}{A vector of daily maximum temperature values of the same
length (and assumed to be the same dates) as \code{daily.precip}.}

\item{average.rainfall}{Reference value for average annual rainfall.}
}
\value{
A vector of KBDI values of the same length as the input precipitation
  and temperature vectors.
}
\description{
Following Finkele et al. (2006), KBDI is calculated as yesterday's KBDI
plus evapo-transpiration minus effective rainfall.
}
