% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_functions.R
\name{calculate_effective_rainfall}
\alias{calculate_effective_rainfall}
\title{Calculate effective rainfall values from daily values}
\usage{
calculate_effective_rainfall(precip.daily)
}
\arguments{
\item{precip.daily}{Vector of daily precipitation values.}
}
\value{
A vector of effective rainfall values.
}
\description{
For KBDI calculation (Finkele et al. 2006), the first 5mm of rain
do not count towards evapo-transpiration. If it rains less than 5mm over several days,
a running balance is kept. Once this balance exceeds 5mm, the first 5mm
is removed for canopy interception/runoff and the remainder
constitutes effective rainfall. Any day without rain resets the balance to zero.
}
