% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_functions.R
\name{calculate_station_ffdi}
\alias{calculate_station_ffdi}
\title{Calculate FFDI for specified weather stations}
\usage{
calculate_station_ffdi(
  db,
  stations,
  av.rainfall.years = 2001:2015,
  min.days.per.year = 300
)
}
\arguments{
\item{db}{A database connection pool object as returned by
\code{\link{bom_db_open}} or \code{\link{bom_db_create}}.}

\item{stations}{A vector of one or more integer weather station identifiers.}
}
\description{
Calculate FFDI for specified weather stations
}
