% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COLUMN_LOOKUP}
\alias{COLUMN_LOOKUP}
\title{Lookup table for station data fields}
\format{
An object of class \code{data.frame} with 49 rows and 4 columns.
}
\usage{
COLUMN_LOOKUP
}
\description{
This data frame is used to map input weather station data columns
to fields in database tables.
}
\keyword{datasets}
