% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{db_creation_script}
\alias{db_creation_script}
\title{Write the database creation script to a file}
\usage{
db_creation_script(dest = "dbcreate.sql")
}
\arguments{
\item{dest}{The destination path and filename. Defaults to
\code{dbcreate.sql} in the current directory.}
}
\description{
The package includes a copy of the SQL script used to create the
PostgreSQL/PostGIS database structure that is assumed by the
package functions. This function simply copies the script into
a specified text file so that you can browse the code or use it
to create your very own database.
}
