% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METADATA_AWS}
\alias{METADATA_AWS}
\title{AWS metadata}
\format{
An object of class \code{data.frame} with 550 rows and 10 columns.
}
\usage{
METADATA_AWS
}
\description{
A data frame that, for each Automaotic Weather Station, gives identifying
number, name (locality), position (latitude and longitude), elevation,
and the month and year for which records commenced.
}
\keyword{datasets}
