% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SQL_CREATE_TABLES}
\alias{SQL_CREATE_TABLES}
\title{SQL statements to create database tables}
\format{
An object of class \code{list} of length 3.
}
\usage{
SQL_CREATE_TABLES
}
\description{
A list of SQL statements used to create the AWS, Synoptic and Stations
tables in a new database.
}
\seealso{
\code{\link{bom_db_create}}
}
\keyword{datasets}
