% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_functions.R
\name{bom_db_calculate_ffdi}
\alias{bom_db_calculate_ffdi}
\title{Calculate FFDI for a given data set.}
\usage{
bom_db_calculate_ffdi(
  dat,
  av.rainfall.years = 2001:2015,
  min.days.per.year = 300,
  av.rainfall.value = NULL
)
}
\arguments{
\item{av.rainfall.years}{A vector of four digit year numbers for the
reference period over which average annual rainfall at each weather station
should be calculated. This is used in the KBDI calculation. Normally the
reference period will be a sequence of contiguous years, but it does not
have to be. Default is \code{2001:2015}. Ignored if a value for
\code{av.rainfall.value} is provided.}

\item{min.days.per.year}{The minimum number of days of rainfall data required
for each year specified by \code{av.rainfall.years}. Default is \code{300}.
Ignored if a value for \code{av.rainfall.value} is provided.}

\item{av.rainfall.value}{A vector of average annual rainfall values to use
instead of calculating averages from weather station data. The vector
should either consist of a single value to use for all weather stations in
the input data, or it should provide a value for each station. The default
(\code{NULL}) means to calculate average annual values from the station
data.}

\item{dat.stn}{A data frame with records for one or more weather stations.
Normally this will made up of records returned from querying the synoptic
or AWS tables in the database. If from another data source, the columns
and column names must conform to those used in the database tables for
variables used to calculate KBDI and FFDI values:
\code{'station', 'year', 'month', 'day', 'hour', 'minute',
      'precipitation', 'temperature', 'relhumidity', 'windspeed'}}
}
\value{
A data frame with columns:
  \code{'station', 'year', 'month', 'day', 'hour', 'minute',}
  \code{'tmaxdaily'} (maximum temperature for calendar day),
  \code{'precipdaily'} (total rainfall from 09:01 to 09:00 the next day),
  \code{'kbdi'} (Keetch-Bryam drought index),
  \code{'drought'} (daily drought factor values),
  \code{'ffdi'} (FFDI value based on daily drought factor and time-step
  values for temperature, wind speed and relative humidity.
}
\description{
This function calculates FFDI (Forest Fire Danger Index)
and the related variables: KBDI (Keetch-Bryam Drought Index) and Drought Factor.
You can call it directly to calculate FFDI for an arbitrary data set of
weather data for one or more stations (see Arguments below for the required
format). The function is also used by \code{bom_db_update_fire} which sets the
values of fire-related variables in the database tables.
}
\examples{
\dontrun{
library(CERMBweather)
library(dplyr, warn.conflicts = FALSE)

# Get synoptic data for selected weather stations
DB <- bom_db_open("/path/to/database/BOM.db")

stns <- c(1006, 1007, 1019, 1020)

dat <- bom_db_synoptic(DB) \%>\%
  filter(station \%in\% stns) \%>\%
  collect()

# Calculate FFDI for stations
ffdi <- bom_db_ffdi(dat)

# Calculate FFDI for stations using non-default range of years
# for average annual rainfall
ffdi <- bom_db_ffdi(dat, av.rainfall.years = 1981:2010)

# Calculate FFDI for stations using a single value for
# average annual rainfall for all weather stations in the
# input data
ffdi <- bom_db_ffdi(dat, av.rainfall.value = 900)

# If you want to relate FFDI and drought factor values to the
# original sub-daily weather values, join to input data
ffdi <- left_join(dat, ffdi,
                  by = c("station", "year", "month", "day", "hour", "minute"))

bom_db_close(DB)
}

}
\seealso{
\code{\link{bom_db_update_fire}}
}
