% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_functions.R
\name{bom_db_calculate_ffdi}
\alias{bom_db_calculate_ffdi}
\title{Calculate FFDI for a given data set.}
\usage{
bom_db_calculate_ffdi(
  dat,
  av.rainfall = NULL,
  datatype = c("guess", "AWS", "Synoptic"),
  windcol = "windspeed"
)
}
\arguments{
\item{dat}{A data frame with records for one or more weather stations.
Normally this will made up of records returned from querying the synoptic
or AWS tables in the database. If from another data source, the columns
and column names must conform to those used in the database tables. The
required columns are:
\code{'station', 'year', 'month', 'day', 'hour', 'minute',
      'precipitation', 'temperature', 'relhumidity'}
plus the specified wind speed column (default: \code{'windspeed'}).}

\item{av.rainfall}{Average annual rainfall value to use in the calculation of
KBDI (on which FFDI relies). Note: if the data include multiple weather
stations, the same average rainfall value is applied to all.}

\item{datatype}{Either a specified data type ('AWS' or 'Synoptic') or 'guess'
(default) to guess the type from the data. Case-insensitive and may be
abbreviated. The data type controls how rainfall values are aggregated to
daily values for the calculation of KBDI and drought factor. AWS rainfall
values are total since 9am, whereas Synoptic values are rainfall for the
record time step. If 'guess', the function checks whether non-zero rainfall
values are always ascending during each 24 hour period to 9am which
indicates AWS data.}

\item{windcol}{Name of the column of wind speed values to use for FFDI
calculation. The default is to use \code{'windspeed'}. AWS data also has
a \code{'windgust'} column (maximum wind gust in previous 10 minutes).}
}
\value{
A data frame with columns:
  \code{'station', 'year', 'month', 'day', 'hour', 'minute',}
  \code{'tmaxdaily'} (maximum temperature for calendar day),
  \code{'precipdaily'} (total rainfall from 09:01 to 09:00 the next day),
  \code{'kbdi'} (Keetch-Bryam drought index),
  \code{'drought'} (daily drought factor values),
  \code{'ffdi'} (FFDI value based on daily drought factor and time-step
    values for temperature, wind speed and relative humidity.
}
\description{
This function calculates FFDI (Forest Fire Danger Index) and the related
variables: KBDI (Keetch-Bryam Drought Index) and Drought Factor. You can call
it directly to calculate FFDI for an arbitrary data set of weather data for
one or more stations (see Arguments below for the required format). The
function is also used by \code{bom_db_update_fire} which sets the values of
fire-related variables in the database tables.
}
\details{
If there are missing days in the time series, dummy records will be added
with all weather variables set to \code{NA}. This allows KBDI, drought factor
and FFDI to be calculated for the rest of the time series. FFDI values for
the dummy records will be set to \code{NA}. KBDI values (and resulting
drought factor values) are set following the protocol described for
\code{\link{bom_db_kbdi}}. The records for each station must cover a period
of at least 21 days for FFDI to be calculated, since Drought Factor requires
an initial 20 days of daily rainfall data.
}
\examples{
\dontrun{
library(CERMBweather)
library(dplyr, warn.conflicts = FALSE)

dat.ffdi <- bom_db_calculate_ffdi(dat.stations, av.rainfall = 800)

# If you want to relate FFDI and drought factor values to the
# original sub-daily weather values, join to input data
dat.ffdi <- left_join(dat.stations, dat.ffdi,
                  by = c("station", "year", "month", "day", "hour", "minute"))
}

}
\seealso{
\code{\link{bom_db_update_fire}}
}
