% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_check_datetimes}
\alias{bom_db_check_datetimes}
\title{Check that records in a data set form an uninterrupted time series}
\usage{
bom_db_check_datetimes(dat, daily)
}
\arguments{
\item{dat}{A data frame with date columns (integer year, month and day),
optional time columns (integer hour and minute) and possibly other
variables. Normally a column of integer station identifiers will be present
but this is optional. If missing, it will be assumed that all records pertain to
a single weather station.}

\item{daily}{If \code{TRUE}, expect only one record per day and return a
failed check result if this is not the case. Note that a value must be
supplied for this argument.}
}
\value{
A nested list with one element per station, where each element is a
  list consisting of:
  \describe{
    \item{station}{Integer station identifier (Set to -1 if no station
      column is provided).}
    \item{ok}{Logical value indicating success or failure of checks.}
    \item{gaps}{Dates (as Date objects) before which each gap in the time
      series occurs.}
  }
}
\description{
Given a data frame of daily or sub-daily records for one or more weather
stations, this function checks whether there are any gaps in the time series.
A gap is defined as one or more missing days. The input data records can be
in any order. It is primarily a helper function, called by other functions in
the package that require an uninterrupted time series, but can also be used
directly.
}
