% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_close}
\alias{bom_db_close}
\title{Close the connection to a database}
\usage{
bom_db_close(db)
}
\arguments{
\item{db}{A database connection pool object as returned by
\code{\link{bom_db_open}} or \code{\link{bom_db_create}}.}
}
\value{
Invisibly returns TRUE if the connection was closed, or FALSE
  otherwise (e.g. the connection had been closed previously).
}
\description{
Given a database connection pool object, as returned by
\code{\link{bom_db_create}} or \code{\link{bom_db_open}}, this function
checks whether the connection is open and, if so, closes it. After being
closed, the connection object can no longer be used.
}
\examples{
\dontrun{
DB <- bom_db_open("c:/foo/bar/weather.db")

# do things with the database, then...

bom_db_close(DB)
}

}
