% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_functions.R
\name{bom_db_drought}
\alias{bom_db_drought}
\title{Calculate daily drought factor values}
\usage{
bom_db_drought(dat.daily, assume.order = FALSE)
}
\arguments{
\item{dat.daily}{A data frame of daily weather data. Must have columns: year,
month, day, precipdaily, kbdi. Optionally, it can also have column:
drought. Additional columns are permitted but will be ignored. See
Description for details.}

\item{assume.order}{If \code{TRUE} the daily data are assumed to be in
calendar order and form an uninterrupted time series. If \code{FALSE}
(default) this will be checked. This option provides for more efficient
processing as part of a pipeline, but should be left at the default when
the function is called directly.}
}
\value{
A data frame with columns: year, month, day, kbdi, drought.
}
\description{
Blah blah blah
}
