% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_functions.R
\name{bom_db_drought}
\alias{bom_db_drought}
\title{Calculate daily drought factor}
\usage{
bom_db_drought(daily.precip, daily.tmax, average.rainfall)
}
\arguments{
\item{daily.precip}{A vector of daily (assumed sequential) precipitation
values.}

\item{daily.tmax}{A vector of daily maximum temperature values of the same
length (and assumed to be the same dates) as \code{daily.precip}.}

\item{average.rainfall}{Reference value for average annual rainfall.}
}
\value{
A data frame with columns kbdi and drought.
}
\description{
Calculates daily drought factor based on the Keetch-Byram drought index
Following Finkele et al. (2006), KBDI is calculated as yesterday's KBDI
plus evapo-transpiration minus effective rainfall.
}
