% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_import}
\alias{bom_db_import}
\title{Import BOM weather station data into a database}
\usage{
bom_db_import(db, datapath, stations = NULL, allow.missing = TRUE)
}
\arguments{
\item{db}{A database connection pool object as returned by
\code{\link{bom_db_open}} or \code{\link{bom_db_create}}.}

\item{datapath}{Character path to one of the following: an individual weather
station data file in CSV format; a directory containing one or more data
files; or a zip file containing one or more such files. If a zip file
(identified by a '.zip' extension) the path should be to a single file.}

\item{stations}{Either NULL (default) to import data for all stations, or a
character vector of station identifiers. Ignored if \code{datapath} is a}

\item{allow.missing}{If TRUE (default) and specific stations were requested,
the function will silently ignore any that are missing in the directory or
zip file. If FALSE, missing stations result in an error.}
}
\value{
\code{TRUE} if the process was completed, or \code{FALSE} otherwise.
  Note that completion of the process means only that the input data were
  successfully read, not necessarily that new records were added to the
  database.
}
\description{
This function can read delimited text data, in the format used by the Bureau
of Meteorology, from an individual weather station file or a directory or zip
file containing one or more such files, stations in CSV format, and import
them into a SQLite database. In the case of the input data source being a
directory or zip file, weather station files are identified by searching for
names that include 'Data' followed by digits and underscores, with the file
extension '.txt'. Any input records already present in the database are
silently ignored. Note that fire-related variables (KBDI, drought factor and
FFDI) are \strong{not} calculated for the new records. call the function
\code{bom_db_update_fire} to do this.
}
\examples{
\dontrun{
# Sometime earlier at start of session
DB <- bom_db_init("c:/foo/bar/weather.db", readonly = FALSE)

# Import data from a BOM zip file containing individuals CSV-format
# data files for weather stations
bom_db_import(DB, "c:/bom/updates/some_aws.zip")

# Import data from CSV-format files in a directory
bom_db_import(DB, "c:/bom/archive/aws_files/")

# Import data from an individual CSV-format weather station file
bom_db_import(DB, "c:/foo/bar/HC06D_Data_068228_999999999515426.txt")

# Do other things, then at end of session...
bom_db_close(DB)
}

}
\seealso{
\code{\link{bom_db_update_fire}}
}
