% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_summary}
\alias{bom_db_summary}
\title{Gets a summary of database contents}
\usage{
bom_db_summary(db, by = c("total", "station"))
}
\arguments{
\item{db}{A database connection pool object as returned by
\code{\link{bom_db_open}} or \code{\link{bom_db_create}}.}

\item{by}{One of 'total' for total records per table, or 'station' for
count of records by weather station.}
}
\value{
A data frame with columns: table; station (if argument 'by' was
  'station'); nrecs.
}
\description{
Gets the count of database records in the AWS and Synoptic tables.
}
