% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_summary}
\alias{bom_db_summary}
\title{Gets a summary of database contents}
\usage{
bom_db_summary(db, by = c("total", "station"), approx = TRUE)
}
\arguments{
\item{db}{A database connection pool object as returned by
\code{\link{bom_db_open}} or \code{\link{bom_db_create}}.}

\item{by}{One of 'total' for total records per table, or 'station' for
count of records by weather station.}

\item{approx}{If TRUE (default) and \code{by == "total"} return an
approximate record count based on the maximum RowID for each table. This is
almost instant whereas a full record count is very slow (minutes rather
than seconds) for large tables. However, it is approximate because RowID is
an automatically incremented variable that does not account for any
previously deleted records. Ignored when \code{by == "station"}.}
}
\value{
A data frame with columns: table; station (if argument 'by' was
  'station'); nrecs.
}
\description{
Gets the count of database records in the AWS and Synoptic tables.
}
