% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_summary}
\alias{bom_db_summary}
\title{Gets a summary of database contents}
\usage{
bom_db_summary(db, by = c("total", "station"), approx = TRUE)
}
\arguments{
\item{db}{A database connection pool object as returned by
\code{\link{bom_db_open}} or \code{\link{bom_db_create}}.}

\item{by}{One of 'total' for total records per table, or 'station' for
count of records by weather station.}

\item{approx}{If TRUE (default) and \code{by == "total"} return an
approximate record count. This is almost instant whereas a full record
count is very slow (minutes rather than seconds) for large tables. However,
it is only approximate and might not reflect recent additions or deletions.
Ignored when \code{by == "station"}.}
}
\value{
A data frame with columns: table; station (if argument 'by' was
  'station'); nrecs.
}
\description{
Gets the count of database records in the AWS and Synoptic tables. It can
take forever to get an exact count of records, especially from a PostgreSQL
database. Setting the \code{approx} argument to \code{TRUE} (default) will
return a very fast, approximate total count. This argument is ignored if the
function is called with \code{by = "station"}.
}
