% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_tbl}
\alias{bom_db_tbl}
\title{Gets a \code{tbl} object for use with dplyr functions}
\usage{
bom_db_tbl(con, tblname)
}
\arguments{
\item{con}{An open database connection as returned by
\code{\link{bom_db_init}}.}

\item{tblname}{One of 'Synoptic' or 'AWS'. May be abbreviated and case is
ignored.}
}
\value{
A \code{tbl} object representing the table to use with dplyr.
}
\description{
This function takes a connection to an open database, along with a table
name ('AWS' or 'Synoptic') and returns a dplyr \code{tbl} object for the
table that you can use with dplyr functions as if it were a data frame.
}
\examples{
\dontrun{
# Connect to a database
con <- bom_db_init("c:/path/to/my/database.db")

# Get a tbl object to use as a proxy for a chosen database table
tsynoptic <- bom_db_tbl(con, "syn")

# Field names in the table
colnames(tsynoptic)

# Use dplyr to find the maximum temperature recorded at each
# weather station

library(dplyr)

dat <- tsynoptic \%>\%
  group_by(station) \%>\%
  summarize(maxtemp = max(temperature, na.rm = TRUE)) \%>\%

  # omit records for stations with no temperature values
  filter(!is.na(maxtemp)) \%>\%

  # tell dplyr to execute this query on the database
  collect()
}

}
