% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{bom_db_tbl}
\alias{bom_db_tbl}
\title{Gets a \code{tbl} object for use with dplyr functions}
\usage{
bom_db_tbl(con, tblname)
}
\arguments{
\item{con}{An open database connection as returned by
\code{\link{bom_db_init}}.}

\item{tblname}{One of 'Synoptic' or 'AWS'. May be abbreviated and case is
ignored.}
}
\value{
A \code{tbl} object that can be to reference this table via dplyr.
}
\description{
Gets a \code{tbl} object for use with dplyr functions
}
