% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STATION_METADATA}
\alias{STATION_METADATA}
\title{Weather station metadata}
\format{
An object of class \code{data.frame} with 17885 rows and 8 columns.
}
\usage{
STATION_METADATA
}
\description{
A data frame that, for each BOM weather station, gives identifying
number, name (locality), start of operation (month and year), state
(including 'ANT' for Antarctica), longitude, and latitude. Also
includes average annual precipitation value for each station based
on the NARCLIM p12 layer.
}
\keyword{datasets}
